+++ files/private-chef-cookbooks/private-chef/attributes/default.rb
@@ -71,9 +71,10 @@ default['private_chef']['rabbitmq']['reindexer_vhost'] = '/reindexer'
 default['private_chef']['rabbitmq']['jobs_vhost'] = '/jobs'
 default['private_chef']['rabbitmq']['jobs_user'] = 'jobs'
 default['private_chef']['rabbitmq']['jobs_password'] = 'workcomplete'
-default['private_chef']['rabbitmq']['actionlog_vhost'] = '/actionlog'
-default['private_chef']['rabbitmq']['actionlog_user'] = 'actionlog'
-default['private_chef']['rabbitmq']['actionlog_password'] = 'changeme'
+default['private_chef']['rabbitmq']['actions_user'] = 'actions'
+default['private_chef']['rabbitmq']['actions_password'] = 'changeme'
+default['private_chef']['rabbitmq']['actions_vhost'] = '/analytics'
+default['private_chef']['rabbitmq']['actions_exchange'] = 'actions'
 default['private_chef']['rabbitmq']['node_ip_address'] = '127.0.0.1'
 default['private_chef']['rabbitmq']['node_port'] = '5672'
 default['private_chef']['rabbitmq']['nodename'] = 'rabbit@localhost'
@@ -161,7 +162,6 @@ default['private_chef']['opscode-erchef']['root_metric_key'] = "chefAPI"
 default['private_chef']['opscode-erchef']['depsolver_worker_count'] = 5
 default['private_chef']['opscode-erchef']['depsolver_timeout'] = 5000
 default['private_chef']['opscode-erchef']['max_request_size'] = 1000000
-default['private_chef']['opscode-erchef']['enable_actionlog'] = false
 
 ###
 # Legacy path (required for cookbok migration)
@@ -450,6 +450,7 @@ default['private_chef']['dark_launch']["private-chef"] = true
 default['private_chef']['dark_launch']["sql_users"] = true
 default['private_chef']['dark_launch']["add_type_and_bag_to_items"] = true
 default['private_chef']['dark_launch']["reporting"] = true
+default['private_chef']['dark_launch']["actions"] = false
 
 ###
 # Opscode Account