+++ files/private-chef-cookbooks/private-chef/recipes/rabbitmq.rb
@@ -84,7 +84,7 @@ if is_data_master?
     retries 10
   end
 
-  [ rabbitmq['vhost'], rabbitmq['reindexer_vhost'], rabbitmq['jobs_vhost'], rabbitmq['actionlog_vhost'] ].each do |vhost|
+  [ rabbitmq['vhost'], rabbitmq['reindexer_vhost'], rabbitmq['jobs_vhost'], rabbitmq['actions_vhost'] ].each do |vhost|
     execute "#{rmq_ctl} add_vhost #{vhost}" do
       user opc_username
       not_if "#{rmq_ctl_chpost} list_vhosts| grep #{vhost}"
@@ -104,9 +104,9 @@ if is_data_master?
     retries 10
   end
 
-  execute "#{rmq_ctl} add_user #{rabbitmq['actionlog_user']} #{rabbitmq['actionlog_password']}" do
+  execute "#{rmq_ctl} add_user #{rabbitmq['actions_user']} #{rabbitmq['actions_password']}" do
     user opc_username
-    not_if "#{rmq_ctl_chpost} list_users |grep #{rabbitmq['actionlog_user']}"
+    not_if "#{rmq_ctl_chpost} list_users |grep #{rabbitmq['actions_user']}"
     retries 10
   end
 
@@ -133,9 +133,15 @@ if is_data_master?
     retries 10
   end
 
-  execute "#{rmq_ctl} set_permissions -p #{rabbitmq['actionlog_vhost']} #{rabbitmq['actionlog_user']} \".*\" \".*\" \".*\"" do
+  execute "#{rmq_ctl} set_permissions -p #{rabbitmq['actions_vhost']} #{rabbitmq['user']} \".*\" \".*\" \".*\"" do
     user opc_username
-    not_if "#{rmq_ctl_chpost} list_user_permissions #{rabbitmq['actionlog_user']}|grep #{rabbitmq['actionlog_vhost']}"
+    not_if "#{rmq_ctl_chpost} list_user_permissions #{rabbitmq['user']}|grep #{rabbitmq['actions_vhost']}"
+    retries 10
+  end
+
+  execute "#{rmq_ctl} set_permissions -p #{rabbitmq['actions_vhost']} #{rabbitmq['actions_user']} \".*\" \".*\" \".*\"" do
+    user opc_username
+    not_if "#{rmq_ctl_chpost} list_user_permissions #{rabbitmq['actions_user']}|grep #{rabbitmq['actions_vhost']}"
     retries 10
   end
 end